code/requirements.txt
# ========================================
# PAQUETES EL CLUB v3.0 - Dependencias de Producción
# ========================================

# FastAPI y servidor ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Cache y tareas en segundo plano
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0

# Email y HTTP
emails==0.6.0
httpx==0.25.2

# Templates
jinja2==3.1.2

# Utilidades
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3

# Monitoreo
prometheus-client==0.19.0
structlog==23.2.0

# Archivos
aiofiles==23.2.1
Pillow==10.1.0
email-validator==2.1.0
